rename service to to nwsc_proxy
