+++ python/nwsc_proxy/test/.coveragerc
