+++ python/nwsc_proxy/test/__init__.py
@@ -1,4 +1,4 @@
-"""Add nwsc_dummy_service top-level module to sys.path"""
+"""Add nwsc_proxy top-level module to sys.path"""
 import os
 import sys
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
